[refactor] Main페이지의 로직을 useMap으로 이동
- Main페이지에 있는 로직들을 useMap으로 이동
- 기존 페이지에 로직이 있으면 이상해서 hook로 이동 + 머지 이후 코드를 맞추기 위함
