+++ src/pages/Main.tsx
@@ -1,13 +1,7 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import styled from '../utils/styles/styled';
 import Sidebar from '../components/Sidebar/Sidebar';
 import Map from '../components/Map/Map';
-import { urlToJson } from '../utils/urlParsing';
-import useWholeStyle from '../hooks/common/useWholeStyle';
-import { WholeStyleActionPayload } from '../store/common/type';
-import { initMap } from '../store/map/action';
-import { useDispatch } from 'react-redux';
-import useMap from '../hooks/map/useMap';
 
 interface MainProps {
   location: {
@@ -27,29 +21,13 @@ const MainWrapper = styled.div`
   height: 100%;
 `;
 
-// http://localhost:3000/map?=poi:landmark:labelText:fill:color:white:end
 function Main({
   location: { search, pathname },
 }: MainProps): React.ReactElement {
-  const { changeStyle } = useWholeStyle();
-  const dispatch = useDispatch();
-  const { mapRef } = useMap();
-
-  const initializeMap = () => {
-    if (search && pathname === '/show') {
-      const states = urlToJson();
-      changeStyle(states as WholeStyleActionPayload);
-    }
-  };
-
-  useEffect((): void => {
-    dispatch(initMap(mapRef, initializeMap));
-  }, [search]);
-
   return (
     <MainWrapper>
       {search && pathname === '/show' ? <></> : <Sidebar />}
-      <Map mapRef={mapRef} />
+      <Map pathname={pathname} />
     </MainWrapper>
   );
 }
